requirements.txt
pytorch_lightning==1.0.8
transformers==4.31.0
Pillow==10.1.0
tqdm==4.65.0
ipdb==0.13.4
numpy==1.26.4
einops==0.3.0
pyarrow==11.0.0
sacred==0.8.5
pandas==1.5.2
timm==0.4.9
ftfy==6.1.1
torchvision==0.16.2
torch==2.1.2
torchtext==0.5.0
PyMatting==1.1.10
wget
